shouldNotLogUriIfIdIsAlreadyMigrated
