+++ publication-model/src/main/java/no/unit/nva/model/contexttypes/Journal.java
@@ -1,28 +1,25 @@
 package no.unit.nva.model.contexttypes;
 
 import static java.util.Objects.isNull;
-import static nva.commons.core.attempt.Try.attempt;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import java.net.URI;
 import java.util.Objects;
+import no.unit.nva.model.contexttypes.utils.MigrateSerialPublicationUtil;
 import no.unit.nva.model.exceptions.InvalidSeriesException;
 import nva.commons.core.JacocoGenerated;
-import nva.commons.core.paths.UriWrapper;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
 public class Journal implements Periodical {
 
     private static final String OLD_PATH_JOURNAL = "journal";
-    private static final String NEW_PATH_SERIAL_PUBLICATION = "serial-publication";
-    private static final int INDEX_FROM_END_CHANNEL_TYPE_PATH_ELEMENT = 2;
     private final URI id;
 
     @JsonCreator
     public Journal(@JsonProperty("id") URI id) {
         validate(id);
-        this.id = migratePath(id);
+        this.id = MigrateSerialPublicationUtil.migratePath(id, OLD_PATH_JOURNAL);
     }
 
     public URI getId() {
@@ -53,22 +50,5 @@ public class Journal implements Periodical {
         if (isNull(stringOfUri) || stringOfUri.isBlank()) {
             throw new InvalidSeriesException(stringOfUri);
         }
-    }
-
-    @Deprecated
-    private static URI replaceOldPath(URI id) {
-        return UriWrapper.fromUri(id)
-                   .replacePathElementByIndexFromEnd(INDEX_FROM_END_CHANNEL_TYPE_PATH_ELEMENT,
-                                                     NEW_PATH_SERIAL_PUBLICATION)
-                   .getUri();
-    }
-
-    @Deprecated
-    private URI migratePath(URI id) {
-        return attempt(() -> UriWrapper.fromUri(id).getPath()
-                                 .getPathElementByIndexFromEnd(INDEX_FROM_END_CHANNEL_TYPE_PATH_ELEMENT)).toOptional()
-                   .filter(path -> path.equals(OLD_PATH_JOURNAL))
-                   .map(uri -> replaceOldPath(id))
-                   .orElse(id);
     }
 }
