+++ publication-model/src/test/java/no/unit/nva/model/contexttypes/MigrateJournalIdTest.java
@@ -6,6 +6,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.not;
+import no.unit.nva.model.contexttypes.utils.MigrateSerialPublicationUtil;
 import nva.commons.core.paths.UriWrapper;
 import nva.commons.logutils.LogUtils;
 import org.junit.jupiter.api.Test;
@@ -26,7 +28,7 @@ public class MigrateJournalIdTest {
     @Test
     void shouldLogUriIfOldIdHasUnexpectedPath() {
         var oldId = constructExampleIdWithPath("unexpected-path");
-        var appender = LogUtils.getTestingAppender(Journal.class);
+        var appender = LogUtils.getTestingAppender(MigrateSerialPublicationUtil.class);
 
         new Journal(oldId);
 
@@ -36,10 +38,20 @@ public class MigrateJournalIdTest {
     @Test
     void shouldLogUriIfOldIdHasUnexpectedForm() {
         var oldId = randomUri();
-        var appender = LogUtils.getTestingAppender(Journal.class);
+        var appender = LogUtils.getTestingAppender(MigrateSerialPublicationUtil.class);
 
         new Journal(oldId);
 
         assertThat(appender.getMessages(), containsString(oldId.toString()));
+    }
+
+    @Test
+    void shouldNotLogUriIfIdIsAlreadyMigrated() {
+        var oldId = constructExampleIdWithPath("serial-publication");
+        var appender = LogUtils.getTestingAppender(MigrateSerialPublicationUtil.class);
+
+        new Journal(oldId);
+
+        assertThat(appender.getMessages(), not(containsString(oldId.toString())));
     }
 }
